# [FLINK-16071][python] Optimize the cost of the get item of the Row