--- a/flink-python/pyflink/fn_execution/coder_impl.py
+++ b/flink-python/pyflink/fn_execution/coder_impl.py
@@ -24,22 +24,22 @@ from apache_beam.coders.coder_impl import StreamCoderImpl
 from pyflink.table import Row
 
 
-class RowCoderImpl(StreamCoderImpl):
+class FlattenRowCoderImpl(StreamCoderImpl):
 
     def __init__(self, field_coders):
         self._field_coders = field_coders
+        self._filed_count = len(field_coders)
 
     def encode_to_stream(self, value, out_stream, nested):
         self.write_null_mask(value, out_stream)
-        for i in range(len(self._field_coders)):
+        for i in range(self._filed_count):
             if value[i] is not None:
                 self._field_coders[i].encode_to_stream(value[i], out_stream, nested)
 
     def decode_from_stream(self, in_stream, nested):
-        null_mask = self.read_null_mask(len(self._field_coders), in_stream)
-        assert len(null_mask) == len(self._field_coders)
-        return Row(*[None if null_mask[idx] else self._field_coders[idx].decode_from_stream(
-            in_stream, nested) for idx in range(0, len(null_mask))])
+        null_mask = self.read_null_mask(self._filed_count, in_stream)
+        return [None if null_mask[idx] else self._field_coders[idx].decode_from_stream(
+            in_stream, nested) for idx in range(0, self._filed_count)]
 
     @staticmethod
     def write_null_mask(value, out_stream):
@@ -76,6 +76,18 @@ class RowCoderImpl(StreamCoderImpl):
                 byte_pos += 1
             field_pos += num_pos
         return null_mask
+
+    def __repr__(self):
+        return 'FlattenRowCoderImpl[%s]' % ', '.join(str(c) for c in self._field_coders)
+
+
+class RowCoderImpl(FlattenRowCoderImpl):
+
+    def __init__(self, field_coders):
+        super(RowCoderImpl, self).__init__(field_coders)
+
+    def decode_from_stream(self, in_stream, nested):
+        return Row(*super(RowCoderImpl, self).decode_from_stream(in_stream, nested))
 
     def __repr__(self):
         return 'RowCoderImpl[%s]' % ', '.join(str(c) for c in self._field_coders)