@dianfu 2020-02-17
The test failure is not related to this PR and there is already one ticket for it: [FLINK-16118](https://issues.apache.org/jira/browse/FLINK-16118)